# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Avora",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/past-papers": {
      "post": {
        "tags": [
          "past-papers"
        ],
        "summary": "Upload Past Paper",
        "operationId": "upload_past_paper_api_v1_past_papers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_past_paper_api_v1_past_papers_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PastPaperOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "past-papers"
        ],
        "summary": "List Past Papers",
        "operationId": "list_past_papers_api_v1_past_papers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PastPaperOut"
                  },
                  "title": "Response List Past Papers Api V1 Past Papers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/past-papers/{past_paper_id}": {
      "get": {
        "tags": [
          "past-papers"
        ],
        "summary": "Past Paper Detail",
        "operationId": "past_paper_detail_api_v1_past_papers__past_paper_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "past_paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Past Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PastPaperDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/past-papers/{past_paper_id}/booklet": {
      "get": {
        "tags": [
          "past-papers"
        ],
        "summary": "Past Paper Booklet",
        "description": "The question paper \u2014 readable by enrolled students so they can sit it.",
        "operationId": "past_paper_booklet_api_v1_past_papers__past_paper_id__booklet_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "past_paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Past Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/past-papers/{past_paper_id}/mark-scheme": {
      "get": {
        "tags": [
          "past-papers"
        ],
        "summary": "Past Paper Mark Scheme",
        "description": "Tutor-only \u2014 handing this to a student would defeat the exercise.",
        "operationId": "past_paper_mark_scheme_api_v1_past_papers__past_paper_id__mark_scheme_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "past_paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Past Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/past-papers/{past_paper_id}/attempts": {
      "post": {
        "tags": [
          "past-papers"
        ],
        "summary": "Log Attempt",
        "description": "A student logging a paper they sat. Creates a Submission, so marking,\nreview and evidence all work exactly as they do for homework.",
        "operationId": "log_attempt_api_v1_past_papers__past_paper_id__attempts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "past_paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Past Paper Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_log_attempt_api_v1_past_papers__past_paper_id__attempts_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PastPaperAttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/past-papers/{past_paper_id}/my-attempt": {
      "get": {
        "tags": [
          "past-papers"
        ],
        "summary": "My Attempt",
        "operationId": "my_attempt_api_v1_past_papers__past_paper_id__my_attempt_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "past_paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Past Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PastPaperAttemptOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response My Attempt Api V1 Past Papers  Past Paper Id  My Attempt Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_log_attempt_api_v1_past_papers__past_paper_id__attempts_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "attempted_at": {
            "type": "string",
            "format": "date",
            "title": "Attempted At"
          },
          "timed": {
            "type": "boolean",
            "title": "Timed",
            "default": false
          },
          "time_taken_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Taken Minutes"
          }
        },
        "type": "object",
        "required": [
          "files",
          "attempted_at"
        ],
        "title": "Body_log_attempt_api_v1_past_papers__past_paper_id__attempts_post"
      },
...
      "Body_upload_past_paper_api_v1_past_papers_post": {
        "properties": {
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "session_label": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Session Label"
          },
          "paper_number": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Paper Number"
          },
          "booklet": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Booklet"
          },
          "mark_scheme": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Mark Scheme"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "session_label",
          "paper_number",
          "booklet",
          "mark_scheme"
        ],
        "title": "Body_upload_past_paper_api_v1_past_papers_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PastPaperAttemptOut": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "past_paper_id": {
            "type": "integer",
            "title": "Past Paper Id"
          },
          "session_label": {
            "type": "string",
            "title": "Session Label"
          },
          "paper_number": {
            "type": "string",
            "title": "Paper Number"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timed": {
            "type": "boolean",
            "title": "Timed"
          },
          "time_taken_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Taken Minutes"
          },
          "attempted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempted At"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "raw_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Marks"
          },
          "max_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Marks"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "past_paper_id",
          "session_label",
          "paper_number",
          "subject_name",
          "status",
          "timed",
          "time_taken_minutes",
          "attempted_at",
          "submitted_at"
        ],
        "title": "PastPaperAttemptOut"
      },
      "PastPaperDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "session_label": {
            "type": "string",
            "title": "Session Label"
          },
          "paper_number": {
            "type": "string",
            "title": "Paper Number"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "booklet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booklet Name"
          },
          "mark_scheme_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mark Scheme Name"
          },
          "extraction_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Error"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 0
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/PastPaperQuestionOut"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_id",
          "session_label",
          "paper_number",
          "total_marks",
          "duration_minutes",
          "booklet_name"
        ],
        "title": "PastPaperDetail"
      },
      "PastPaperOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "session_label": {
            "type": "string",
            "title": "Session Label"
          },
          "paper_number": {
            "type": "string",
            "title": "Paper Number"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "booklet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booklet Name"
          },
          "mark_scheme_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mark Scheme Name"
          },
          "extraction_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Error"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_id",
          "session_label",
          "paper_number",
          "total_marks",
          "duration_minutes",
          "booklet_name"
        ],
        "title": "PastPaperOut"
      },
      "PastPaperQuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "text_summary": {
            "type": "string",
            "title": "Text Summary"
          },
          "max_marks": {
            "type": "integer",
            "title": "Max Marks"
          },
          "has_mark_scheme": {
            "type": "boolean",
            "title": "Has Mark Scheme"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "text_summary",
          "max_marks",
          "has_mark_scheme"
        ],
        "title": "PastPaperQuestionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}